Changed the styles and added Andy's new classes for his forthcoming "thing". Altered the timezone settings in prefs and the new registration page to do timezones correctly (some regions of the world are not an whole number of hours ahead or behind GMT/UTC). Added a print option to the bottom of posts and a Print button to display.php.

+++ docs/schema.sql
@@ -444,7 +444,7 @@ CREATE TABLE USER_PREFS (
   HOMEPAGE_URL varchar(255) default NULL,
   PIC_URL varchar(255) default NULL,
   EMAIL_NOTIFY char(1) default NULL,
-  TIMEZONE tinyint(4) default NULL,
+  TIMEZONE decimal(2,1) default NULL,
   DL_SAVING char(1) default NULL,
   MARK_AS_OF_INT char(1) default NULL,
   POSTS_PER_PAGE tinyint(3) unsigned default NULL,
